Abstracted TCP specific flow handling.
