+++ nmeta2dpae/flow/tcp_flow.py
@@ -16,12 +16,15 @@
 """
 This module is part of the nmeta2 suite
 .
-It provides an abstraction for a TCP flow that links to
-a MongoDB database and changes to the context of the flow
-that a supplied packet belongs to
+It provides an abstraction for a TCP flow that links to a MongoDB
+database and changes to the context of the flow that a supplied packet
+belongs to. We implement the Flow class.
 .
 Version 2.x Toulouse Code
 """
+
+# Inherit from parent class
+import flow
 
 #*** For packet methods:
 import socket
@@ -32,22 +35,13 @@ import dpkt
 #*** mongodb Database Import:
 from pymongo import MongoClient
 
-#*** For hashing flow 5-tuples:
-import hashlib
-
 # Class variables containing protocol type values to reduce the amount
 # of referencing to modules.
 _ETH_TYPE_IP = dpkt.ethernet.ETH_TYPE_IP
-_ETH_TYPE_ARP = dpkt.ethernet.ETH_TYPE_ARP
-_ETH_TYPE_VLAN = dpkt.ethernet.ETH_TYPE_8021Q
-_ETH_TYPE_IP6 = dpkt.ethernet.ETH_TYPE_IP6
-_ETH_TYPE_LLDP = dpkt.ethernet.ETH_TYPE_LLDP
-_IP_PROTO_ICMP = dpkt.ip.IP_PROTO_ICMP
 _IP_PROTO_TCP = dpkt.ip.IP_PROTO_TCP
-_IP_PROTO_UDP = dpkt.ip.IP_PROTO_UDP
 
 
-class Flow(object):
+class TCPFlow(flow.Flow):
     """
     An object that represents a flow that we are classifying
 
@@ -92,8 +86,8 @@ class Flow(object):
         flow.packet_count   # Unique packets registered for the flow
         flow.client         # The IP that is the originator of the TCP
                             #  session (if known, otherwise 0)
-        flow.server         # The IP that is the destination of the TCP session
-                            #  session (if known, otherwise 0)
+        flow.server         # The IP that is the destination of the TCP
+                            # session (if known, otherwise 0)
 
     Methods available for Classifiers (assumes class instantiated as
     an object called 'flow'):
@@ -108,37 +102,24 @@ class Flow(object):
     """
 
     def __init__(self, logger, mongo_addr, mongo_port):
+        """Initialise an instance of the TCPFlow class for a new flow.
+
+        Passed layer 3/4 parameters. Add an entry to the FCIP database
+        if it doesn't already exist. If it does exist, update it.
+
+        This is specific to TCP.
         """
-        Initialise an instance of the Flow class for a new
-        flow. Passed layer 3/4 parameters.
-        Add an entry to the FCIP database if it doesn't
-        already exist. If it does exist, update it.
-        Only works for TCP at this stage.
-        """
+        flow.Flow.__init__(self)
         self.logger = logger
         #*** Maximum packets in a flow before finalising:
         self.max_packet_count = 10
 
-        #*** Initialise packet variables:
-        self.ip_src = 0
-        self.ip_dst = 0
+        #*** Initialise specifc packet variables:
         self.tcp_src = 0
         self.tcp_dst = 0
         self.tcp_flags = 0
         self.tcp_seq = 0
         self.tcp_acq = 0
-        self.payload = 0
-
-        #*** Initialise flow variables:
-        self.finalised = 0
-        self.packet_length = 0
-        self.packet_count = 0
-        self.fcip_doc = {}
-        self.fcip_hash = 0
-        self.client = 0
-        self.server = 0
-        self.packet_direction = 'unknown'
-        self.suppressed = 0
 
         #*** Start mongodb:
         self.logger.info("Connecting to mongodb database...")
@@ -147,7 +128,7 @@ class Flow(object):
         #*** Connect to specific databases and collections in mongodb:
         #*** FCIP (Flow Classification in Progress) database:
         db_fcip = mongo_client.fcip_database
-        self.fcip = db_fcip.fcip
+        self.fcip = db_fcip.fcip # MAKE TCP SPECIFIC
 
         #*** DPAE database - delete all previous entries:
         result = self.fcip.delete_many({})
@@ -167,8 +148,8 @@ class Flow(object):
         self.packet_length = len(pkt)
         #*** Read into dpkt:
         eth = dpkt.ethernet.Ethernet(pkt)
-        eth_src = _mac_addr(eth.src)
-        eth_dst = _mac_addr(eth.dst)
+        eth_src = flow.mac_addr(eth.src)
+        eth_dst = flow.mac_addr(eth.dst)
         eth_type = eth.type
         #*** We only support IPv4 (TBD: add IPv6 support):
         if eth_type != _ETH_TYPE_IP:
@@ -177,7 +158,7 @@ class Flow(object):
         ip = eth.data
         self.ip_src = socket.inet_ntop(socket.AF_INET, ip.src)
         self.ip_dst = socket.inet_ntop(socket.AF_INET, ip.dst)
-        # TODO divide up for UDP
+
         #*** We only support TCP:
         if ip.p != _IP_PROTO_TCP:
             self.logger.error("Non TCP packet, ip_proto=%s",
@@ -192,8 +173,8 @@ class Flow(object):
         self.tcp_flags = tcp.flags
         self.payload = tcp.data
         #*** Generate a hash unique to flow for packets in either direction
-        self.fcip_hash = _hash_5tuple(self.ip_src, self.ip_dst, self.tcp_src,
-                                        self.tcp_dst, proto)
+        self.fcip_hash = flow.hash_5tuple(self.ip_src, self.ip_dst,
+                                          self.tcp_src, self.tcp_dst, proto)
         #*** Check to see if we already know this identity:
         db_data = {'hash': self.fcip_hash}
         self.fcip_doc = self.fcip.find_one(db_data)
@@ -275,98 +256,6 @@ class Flow(object):
             self.logger.debug("min_interpacket_interval is %s",
                                             self.min_interpacket_interval())
 
-    def max_packet_size(self):
-        """
-        Return the size of the largest packet in the flow (in either direction)
-        """
-        return max(self.fcip_doc['packet_lengths'])
-
-    def max_interpacket_interval(self):
-        """
-        Return the size of the largest inter-packet time interval
-        in the flow (assessed per direction in flow).
-        .
-        Note: slightly inaccurate due to floating point rounding.
-        """
-        max_c2s = 0
-        max_s2c = 0
-        count_c2s = 0
-        count_s2c = 0
-        prev_c2s_idx = 0
-        prev_s2c_idx = 0
-        for idx, direction in enumerate(self.fcip_doc['packet_directions']):
-            if direction == 'c2s':
-                count_c2s += 1
-                if count_c2s > 1:
-                    current_ts = self.fcip_doc['packet_timestamps'][idx]
-                    prev_ts = self.fcip_doc['packet_timestamps'][prev_c2s_idx]
-                    delta = current_ts - prev_ts
-                    if delta > max_c2s:
-                        max_c2s = delta
-                    prev_c2s_idx = idx
-            elif direction == 's2c':
-                count_s2c += 1
-                if count_s2c > 1:
-                    current_ts = self.fcip_doc['packet_timestamps'][idx]
-                    prev_ts = self.fcip_doc['packet_timestamps'][prev_s2c_idx]
-                    delta = current_ts - prev_ts
-                    if delta > max_s2c:
-                        max_s2c = delta
-                    prev_s2c_idx = idx
-            else:
-                #*** Don't know direction so ignore:
-                pass
-        #*** Return the largest interpacket delay overall:
-        if max_c2s > max_s2c:
-            return max_c2s
-        else:
-            return max_s2c
-
-    def min_interpacket_interval(self):
-        """
-        Return the size of the smallest inter-packet time interval
-        in the flow (assessed per direction in flow)
-        .
-        Note: slightly inaccurate due to floating point rounding.
-        """
-        min_c2s = 0
-        min_s2c = 0
-        count_c2s = 0
-        count_s2c = 0
-        prev_c2s_idx = 0
-        prev_s2c_idx = 0
-        for idx, direction in enumerate(self.fcip_doc['packet_directions']):
-            if direction == 'c2s':
-                count_c2s += 1
-                if count_c2s > 1:
-                    current_ts = self.fcip_doc['packet_timestamps'][idx]
-                    prev_ts = self.fcip_doc['packet_timestamps'][prev_c2s_idx]
-                    delta = current_ts - prev_ts
-                    if not min_c2s or delta < min_c2s:
-                        min_c2s = delta
-                    prev_c2s_idx = idx
-            elif direction == 's2c':
-                count_s2c += 1
-                if count_s2c > 1:
-                    current_ts = self.fcip_doc['packet_timestamps'][idx]
-                    prev_ts = self.fcip_doc['packet_timestamps'][prev_s2c_idx]
-                    delta = current_ts - prev_ts
-                    if not min_s2c or delta < min_s2c:
-                        min_s2c = delta
-                    prev_s2c_idx = idx
-            else:
-                #*** Don't know direction so ignore:
-                pass
-        #*** Return the smallest interpacket delay overall, watch out for
-        #***  where we didn't get a calculation (don't return 0 unless both 0):
-        if not min_s2c:
-            #*** min_s2c not set so return min_c2s as it might be:
-            return min_c2s
-        elif 0 < min_c2s < min_s2c:
-            return min_c2s
-        else:
-            return min_s2c
-
     def set_suppress_flow(self):
         """
         Set the suppressed attribute in the flow database
@@ -437,33 +326,3 @@ def _is_tcp_syn(tcp_flags):
         return 1
     else:
         return 0
-
-def _hash_5tuple(ip_A, ip_B, tp_src, tp_dst, proto):
-    """
-    Generate a predictable hash for the 5-tuple which is the
-    same not matter which direction the traffic is travelling
-    """
-    if ip_A > ip_B:
-        direction = 1
-    elif ip_B > ip_A:
-        direction = 2
-    elif tp_src > tp_dst:
-        direction = 1
-    elif tp_dst > tp_src:
-        direction = 2
-    else:
-        direction = 1
-    hash_5t = hashlib.md5()
-    if direction == 1:
-        flow_tuple = (ip_A, ip_B, tp_src, tp_dst, proto)
-    else:
-        flow_tuple = (ip_B, ip_A, tp_dst, tp_src, proto)
-    flow_tuple_as_string = str(flow_tuple)
-    hash_5t.update(flow_tuple_as_string)
-    return hash_5t.hexdigest()
-
-def _mac_addr(address):
-    """
-    Convert a MAC address to a readable/printable string
-    """
-    return ':'.join('%02x' % ord(b) for b in address)
